Clear vehicle identifier on disconnect
